added reusable controllers
