+++ data/db.json
@@ -13,28 +13,10 @@
       "status": "done"
     },
     {
-      "id": "16c9",
-      "title": "fix ticket #1203",
-      "priorityLevel": "medium",
-      "status": "ongoing"
-    },
-    {
       "id": "f86a",
       "title": "create landing page test",
       "priorityLevel": "high",
       "status": "ongoing"
-    },
-    {
-      "id": "164c",
-      "title": "test 23",
-      "priorityLevel": "low",
-      "status": "done"
-    },
-    {
-      "id": "458f",
-      "title": "test1",
-      "priorityLevel": "low",
-      "status": "todo"
     },
     {
       "id": "f166",
@@ -47,6 +29,30 @@
       "title": "zcvzx",
       "priorityLevel": "medium",
       "status": "todo"
+    },
+    {
+      "id": "09c2",
+      "title": "fix ticket #4120",
+      "priorityLevel": "high",
+      "status": "todo"
+    },
+    {
+      "id": "a592",
+      "title": "Fix responsive for mobile",
+      "priorityLevel": "high",
+      "status": "ongoing"
+    },
+    {
+      "id": "a592",
+      "title": "Fix responsive for mobile",
+      "priorityLevel": "medium",
+      "status": "ongoing"
+    },
+    {
+      "id": "e9e2",
+      "title": "submit test 1",
+      "status": "done",
+      "priorityLevel": "low"
     }
   ]
 }
